se-interview-project/se-interview/requirements.txt

viam-sdk==0.33.0

typing-extensions
